+++ Packages/com.github.kurotu.vrc-quest-tools/Runtime/Components/AvatarConverterSettings.cs
@@ -81,7 +81,7 @@ namespace KRT.VRCQuestTools.Components
         /// NDMF phase to convert the avatar.
         /// </summary>
         [SerializeField]
-        public AvatarConverterNdmfPhase ndmfPhase = AvatarConverterNdmfPhase.Transforming;
+        public AvatarConverterNdmfPhase ndmfPhase = AvatarConverterNdmfPhase.Auto;
 
         /// <summary>
         /// Gets avatar descriptor of the avatar root object.