+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- [NDMF] Add `Auto` option to NDMF Phase in `VQT Avatar Converter Settings` and `VQT Material Conversion Settings` components.
+    - This option automatically sets the phase based on the current project.
+        - Default: Optimizing Phase
+        - VRCFury exists: Transforming Phase
+
+### Changed
+- [NDMF] Default NDMF Phase in `VQT Avatar Converter Settings` and `VQT Material Conversion Settings` components is changed to `Auto`.
+
 ## [2.10.0-beta.3] - 2025-05-24
 
 ### Added